ml_pipeline/requirements.txt
pandas==2.1.4
beautifulsoup4==4.12.3
pyyaml==6.0.1
requests==2.31.0
torch==2.1.2
transformers==4.36.2
datasets==2.16.1
accelerate==0.25.0
peft==0.7.1
bitsandbytes==0.41.3
trl==0.7.9
scipy==1.11.4
